allow reading JSON files
Allow input to be JSON files, such that it is easier to procedurally generate
the files (e.g. from another script that outputs JSON). JSON input will be
read as a dictionary and put into a namespace.

Improve the README by mentioning the two input formats, and clean up the
example given such that it is easier to read and does not rely on the VMAF
configuration (which is not necessary to run the model).

Signed-off-by: Werner Robitza <werner.robitza@gmail.com>